Update verification environment to support MSPGCC Uniarch (based on GCC 4.5 and later)
git-svn-id: http://opencores.org/ocsvn/openmsp430/openmsp430/trunk@141 c36ddccb-d6af-41aa-abd5-588c0c1d0201

+++ core/sim/rtl_sim/bin/asm2ihex.sh
@@ -37,11 +37,11 @@
 ###############################################################################
 #                            Parameter Check                                  #
 ###############################################################################
-EXPECTED_ARGS=6
+EXPECTED_ARGS=7
 if [ $# -ne $EXPECTED_ARGS ]; then
   echo "ERROR    : wrong number of arguments"
-  echo "USAGE    : asm2ihex.sh <test name> <test assembler file> <linker script>   <prog mem size> <data mem size> <peripheral addr space size>"
-  echo "Example  : asm2ihex.sh c-jump_jge  ../src/c-jump_jge.s43 ../bin/template.x 2048            128             512"
+  echo "USAGE    : asm2ihex.sh <test name> <test assembler file> <linker script> <assembler define>  <prog mem size> <data mem size> <peripheral addr space size>"
+  echo "Example  : asm2ihex.sh c-jump_jge  ../src/c-jump_jge.s43 ../bin/template.x ../bin/pmem.h 2048            128             512"
   exit 1
 fi
 
@@ -58,24 +58,32 @@ if [ ! -e $3 ]; then
     echo "Linker definition file template doesn't exist: $3"
     exit 1
 fi
+if [ ! -e $4 ]; then
+    echo "Assembler definition file template doesn't exist: $4"
+    exit 1
+fi
 
 
 ###############################################################################
 #               Generate the linker definition file                           #
 ###############################################################################
 
-PER_SIZE=$6
-DMEM_SIZE=$5
-PMEM_SIZE=$4
+PER_SIZE=$7
+DMEM_SIZE=$6
+PMEM_SIZE=$5
 PMEM_BASE=$((0x10000-$PMEM_SIZE))
 STACK_INIT=$((PER_SIZE+0x0080))
 
 cp  $3  ./pmem.x
+cp  $4  ./pmem_defs.asm
 sed -i "s/PMEM_BASE/$PMEM_BASE/g"    pmem.x
 sed -i "s/PMEM_SIZE/$PMEM_SIZE/g"    pmem.x
 sed -i "s/DMEM_SIZE/$DMEM_SIZE/g"    pmem.x
 sed -i "s/PER_SIZE/$PER_SIZE/g"      pmem.x
 sed -i "s/STACK_INIT/$STACK_INIT/g"  pmem.x
+
+sed -i "s/PER_SIZE/$PER_SIZE/g"      pmem_defs.asm
+sed -i "s/PMEM_SIZE/$PMEM_SIZE/g"    pmem_defs.asm
 
 
 ###############################################################################
